Close service provider files when done reading them
scala.io.Source doesn't implicitly close the file when the iterator
returned by getLines reaches the end.  Since it assumes implicitly
the whole contents of the file fit into memory anyway, I've just
replaced it with IO.readLines.

Also, service provider files are required to be encoded in UTF-8, so
I've specified that too instead of leaving it up to the whims of default
charsets.