Make readme less condencending.
The old readme said "If you need to tell sbt-assembly to ignore JARs, you're
probably doing it wrong" and then followed up with "If you Spark people
want to [do it anyways]", which I think is unnecessarily condensending.
The new readme tones it down by saying "If you're using Spark and want to ...".